Enemy sprite working, added damage to enemy with collision occurs
If the user runs into an enemy, the player takes 10f damage and if the players health reaches 0, The player despawns
