+++ Assets/Assets/Sprites/Red.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a277b698db69c774e8f8633dbce4d773
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
